вынесла компоненты для таблицы одельно
